warning fix
경고 오류 수정 및 utf-8 세팅

+++ win32_socket_multi_chatting/win32_socket_multi_chatting/Client.h
@@ -26,7 +26,7 @@ typedef struct
 {
 	int price;		// 가격
 	char productName[PRODUCT_NAME_SIZE];	// 상품명
-	float validity;	// 유효기간
+	unsigned int validity;	// 유효기간
 } giftData_t;
 
 enum SendMessageKind
